break more out into @runt/ai
