+++ packages/lib/src/types.ts
@@ -12,6 +12,7 @@ import type {
   RichOutputData,
   schema,
 } from "@runt/schema";
+import type { Logger } from "./logging.ts";
 
 /**
  * Configuration options for runtime agents
@@ -92,6 +93,8 @@ export interface ExecutionContext {
   error: (ename: string, evalue: string, traceback: string[]) => void;
   /** Clear all previous outputs for this cell */
   clear: () => void;
+
+  logger: Logger;
 }
 
 /**
